Module Tjr_pcache__.Pcache_intf

type buf = Tjr_fs_shared.ba_buf

type buf_ops = {

create : int -> buf;

get : int -> buf -> char;

}

val buf_ops : buf_ops

type ('k, 'v, 'r, 'kvop_map) detach_info = {

root_ptr : 'r;

past_map : 'kvop_map;

current_ptr : 'r;

current_map : 'kvop_map;

}

The result of "detaching" the map. We get the abstract map for all but the current node, and information about the current node.

type ('k, 'v, 'r, 'kvop_map, 't) pcache_ops = {

find : 'k -> ('v option, 't) Tjr_monad.m;

insert : 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'k -> (unit, 't) Tjr_monad.m;

detach : unit -> (('k, 'v, 'r, 'kvop_map) detach_info, 't) Tjr_monad.m;

blk_len : unit -> (int, 't) Tjr_monad.m;

pcache_write : unit -> (unit, 't) Tjr_monad.m;

pcache_sync : unit -> (unit, 't) Tjr_monad.m;

}

For the detach operation, we get the map upto the current node, and the map for the current node. NOTE if root_ptr = current_ptr, then nothing was detached.

type ('r, 'kvop_map) pcache_state = {

root_ptr : 'r;

past_map : 'kvop_map;

current_ptr : 'r;

current_map : 'kvop_map;

buf : buf;

buf_pos : int;

next_ptr : 'r option;

blk_len : int;

dirty : bool;

}

val empty_pcache_state : root_ptr:'a -> current_ptr:'a -> empty:'b -> ('a, 'b) pcache_state

Pcache factory

type ('k, 'v, 'r, 'buf, 'kvop_map, 't) pcache_factory_1 = < read_pcache : 'r -> (< tl : (('k, 'v) Tjr_fs_shared.Kvop.kvop list * 'r option) list; hd : 'r * 'buf * int; >, 't) Tjr_monad.m; read_initial_pcache_state : 'r -> (('r, 'kvop_map) pcache_state, 't) Tjr_monad.m; make_pcache_ops : < with_state : (('r, 'kvop_map) pcache_state, 't) Tjr_monad.with_state -> ('k, 'v, 'r, 'kvop_map, 't) pcache_ops; from_root : 'r -> (< with_state : (('r, 'kvop_map) pcache_state, 't) Tjr_monad.with_state; pcache_ops : ('k, 'v, 'r, 'kvop_map, 't) pcache_ops; >, 't) Tjr_monad.m; >; >

type ('k, 'v, 'r, 'buf, 'kvop_map, 't) pcache_factory = < kvop_map_ops : ('k, ('k, 'v) Tjr_fs_shared.Kvop.kvop, 'kvop_map) Tjr_lib.Tjr_map.map_ops; empty_pcache_state : root_ptr:'r -> current_ptr:'r -> ('r, 'kvop_map) pcache_state; with_read_blk_as_buf : read_blk_as_buf:('r -> ('buf, 't) Tjr_monad.m) -> flush_tl:(('r, 'kvop_map) pcache_state -> (unit, 't) Tjr_monad.m) -> ('k, 'v, 'r, 'buf, 'kvop_map, 't) pcache_factory_1; with_blk_dev_ops : blk_dev_ops:('r, 'buf, 't) Tjr_fs_shared.blk_dev_ops -> ('k, 'v, 'r, 'buf, 'kvop_map, 't) pcache_factory_1; >