Module Pcache_intf.Pcache_ops

type ('k, 'v, 'r, 'kvop_map, 't) pcache_ops = {

find : 'k -> ('v option, 't) Tjr_monad.m;

insert : 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'k -> (unit, 't) Tjr_monad.m;

detach : unit -> (('k, 'v, 'r, 'kvop_map) detach_info, 't) Tjr_monad.m;

blk_len : unit -> (int, 't) Tjr_monad.m;

get_origin : unit -> ('r Tjr_plist.Pl_origin.t, 't) Tjr_monad.m;

pcache_sync : unit -> (unit, 't) Tjr_monad.m;

}

For the detach operation, we get the map upto the current node, and the map for the current node. NOTE if root_ptr = current_ptr, then nothing was detached.