Module Tjr_pcache__.Pvt_make

module type S = sig ... end

Make functor argument type, including types and values

module Chk : sig ... end

For testing

type ('a, 'b, 'c, 'd, 'e, 'f, 'g) c_dmap = {

monad_ops : 'a;

marshalling_config : 'b;

kvop_map_ops : 'c;

blk_ops : 'd;

blk_alloc : 'e;

with_dmap : 'f;

write_to_disk : 'g;

}

FIXME write_to_disk is more like "sync to disk", and is intended to sync just the current tl

module Make : functor (S : S) -> sig ... end

module Make_with_fixed_types : functor (S : sig ... end) -> sig ... end

Make alternative based on fixing just the types, not the values; halfway house between functor and function

val make : ('t Tjr_monad.monad_ops, ('k, 'v, 'r) Tjr_pcache.Pcache_intf.marshalling_config, ('k, ('k, 'v) Tjr_fs_shared.kvop, 'kvop_map) Tjr_pcache.Pcache_intf.Pvt.pcache_map_ops, 'blk Tjr_fs_shared.blk_ops, unit -> ('r, 't) Tjr_monad.m, (('r, 'kvop_map) Tjr_pcache.Pcache_intf.Pvt.dmap_state, 't) Tjr_monad.with_state, ('r, 'kvop_map) Tjr_pcache.Pcache_intf.Pvt.dmap_state -> (unit, 't) Tjr_monad.m) c_dmap -> ('k, 'v, 'r, 'kvop_map, 't) Tjr_pcache.Pcache_intf.dmap_ops

make as a function