Module Tjr_pcache__Pcache_intf

type buf = Tjr_fs_shared.ba_buf

type buf_ops = {

create : int -> buf;

get : int -> buf -> char;

}

val buf_ops : buf_ops

type ('k, 'v, 'r, 'kvop_map) detach_info = {

root_ptr : 'r;

past_map : 'kvop_map;

current_ptr : 'r;

current_map : 'kvop_map;

}

The result of "detaching" the map. We get the abstract map for all but the current node, and information about the current node.

type ('k, 'v, 'r, 'kvop_map, 't) dmap_ops = {

find : 'k -> ('v option, 't) Tjr_monad.m;

insert : 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'k -> (unit, 't) Tjr_monad.m;

detach : unit -> (('k, 'v, 'r, 'kvop_map) detach_info, 't) Tjr_monad.m;

block_list_length : unit -> (int, 't) Tjr_monad.m;

dmap_write : unit -> (unit, 't) Tjr_monad.m;

dmap_sync : unit -> (unit, 't) Tjr_monad.m;

read_pcache : root:'r -> read_blk_as_buf:('r -> (buf, 't) Tjr_monad.m) -> ((('k, 'v) Tjr_fs_shared.Kvop.kvop list * 'r option) list, 't) Tjr_monad.m;

}

For the detach operation, we get the map upto the current node, and the map for the current node. NOTE if root_ptr = current_ptr, then nothing was detached.

module Pvt : sig ... end

module type MRSHL = sig ... end

type ('k, 'v, 'r) marshalling_config = (module MRSHL with type k = 'k and type r = 'r and type v = 'v)