Module Tjr_pcache__Pcache_intf.Detach_info

type ('k, 'v, 'r, 'kvop_map) detach_info = {

root_ptr : 'r;

past_map : 'kvop_map;

current_ptr : 'r;

current_map : 'kvop_map;

}

The result of "detaching" the map. We get the abstract map for all but the current node, and information about the current node.

type ('k, 'v, 'r, 'kvop_map) t = ('k, 'v, 'r, 'kvop_map) detach_info